requirements.txt
matplotlib
numpy
pytorch
statsmodels